% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_village.R
\name{baseline_donor}
\alias{baseline_donor}
\title{Sets baseline donor}
\usage{
baseline_donor(village)
}
\arguments{
\item{village}{A Village object.}
}
\value{
An updated village object.
}
\description{
Internal generic function that either sets a recommended default or user provided baseline donor. Function also recommends possible alternative donors that have median proliferation rates in the village experiment.
}
\keyword{internal}
