% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_townlet.R
\name{def_priors}
\alias{def_priors}
\title{Define model priors}
\usage{
def_priors(village)
}
\arguments{
\item{village}{A Village object.}
}
\value{
An updated village object.
}
\description{
Internal generic function that sets model priors. Dispersion variances are set by applying an empirical Bayes approach.
}
\keyword{internal}
