% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_townlet.R
\name{extract_eta}
\alias{extract_eta}
\title{Extract draws for eta}
\usage{
extract_eta(village)
}
\arguments{
\item{village}{A Village object.}
}
\value{
A list of draws data frames (for control (beta) and optional treatment (tau_d) donor specific parameters).
}
\description{
Internal generic function that extracts donor specific proliferation effect parameter draws for plotting and significance testing.
}
\keyword{internal}
