% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_townlet.R
\name{extractbeta}
\alias{extractbeta}
\title{Extract draws}
\usage{
extractbeta(village)
}
\arguments{
\item{village}{A Village object.}
}
\value{
A list of draws data frames (for control and optional treatment donor specific parameters).
}
\description{
Internal generic function that extracts donor specific parameter draws for plotting and significance testing.
}
\keyword{internal}
