% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_townlet.R
\name{growthmetric}
\alias{growthmetric}
\title{Donor proliferation effects}
\usage{
growthmetric(village)
}
\arguments{
\item{village}{A Village object.}
}
\value{
An updated village object.
}
\description{
Internal generic function that calculates total donor proliferation effects.
}
\keyword{internal}
