% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_village.R
\name{init_village}
\alias{init_village}
\title{Initiate village experiment data}
\usage{
init_village(
  datapath,
  model = NULL,
  normalize = FALSE,
  outdir = "./",
  name = "test_townlet",
  cls = NULL,
  T0_cutoff = NULL,
  alldonors = TRUE,
  sim = FALSE,
  timeunit = "days",
  baseline = NULL,
  ebayes = TRUE
)
}
\arguments{
\item{datapath}{File path to census composition data (see github tutorial for df format)}

\item{model}{Regression model including all optional donor covariates and treatments (e.g. ~sex + treatment_lead + sex:treatment_lead). If no covariates/treatments included in analysis then model = NULL.}

\item{normalize}{Force normalization of representation so that it sums to 1.}

\item{outdir}{Output directory to save results}

\item{name}{Name for output directory/files}

\item{cls}{Village donor colors to use in plots}

\item{T0_cutoff}{Threshold for excluding under represented donors at T0.}

\item{alldonors}{Include all donors? If FALSE, normalization will be forced.}

\item{sim}{If TRUE, runs in simulation mode (no plots generated).}

\item{timeunit}{Time unit for samples (e.g. days, cell passages, etc.)}

\item{baseline}{Choose baseline donor (see documentation before overriding default)}

\item{ebayes}{Use empirical Bayes method to set dispersion prior variance}
}
\value{
A Village object with formatted data for townlet donor proliferation analysis
}
\description{
Formats user village composition data and checks that required columns are present to run townlet analysis. Plots raw data for inspection and sets baseline donor.
}
\examples{
\dontrun{
test_village <- init_village(
  datapath = "./village_census.csv",
  model = "~sex + treatment_lead + sex:treatment_lead",
  name = "test_townlet",
  outdir = "./"
)
}
}
