% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_townlet.R
\name{model_inputs}
\alias{model_inputs}
\title{Define model inputs}
\usage{
model_inputs(village)
}
\arguments{
\item{village}{A Village object.}
}
\value{
An updated village object.
}
\description{
Internal generic function to define townlet model inputs based on user provided data and model specifications.
}
\keyword{internal}
