% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_townlet.R
\name{modelversion}
\alias{modelversion}
\title{Define townlet model}
\usage{
modelversion(village)
}
\arguments{
\item{village}{A Village object.}
}
\value{
An updated village object.
}
\description{
Internal generic function that selects the appropriate townlet model version based on user provided model specifications.
}
\keyword{internal}
