% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_betas}
\alias{plot_betas}
\title{Plot donor beta parameters per treatment}
\usage{
plot_betas(village, lowci, highci, df_all, df_donor)
}
\arguments{
\item{village}{A Village object.}

\item{lowci}{Lower bound of credible interval}

\item{highci}{Higher bound of credible interval}

\item{df_all}{Data frame of draws.}

\item{df_donor}{Donor identification info}
}
\value{
None
}
\description{
Internal generic function to plot violin plot of beta parameters per treatment.
}
\keyword{internal}
