% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_donorcov}
\alias{plot_donorcov}
\title{Plot donor covariates}
\usage{
plot_donorcov(village, lowci, highci)
}
\arguments{
\item{village}{A Village object.}

\item{lowci}{Lower bound of credible interval}

\item{highci}{Higher bound of credible interval}
}
\value{
List containing plot and data.
}
\description{
Internal generic function to plot violin plot of donor covariates.
}
\keyword{internal}
