% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_ESS}
\alias{plot_ESS}
\title{Plot effective sample size}
\usage{
plot_ESS(village)
}
\arguments{
\item{village}{A Village object.}
}
\value{
None.
}
\description{
Internal generic function to plot effective sample size of all parameters.
}
\keyword{internal}
