% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_eta}
\alias{plot_eta}
\title{Plot total proliferation effect}
\usage{
plot_eta(village, lowci, highci, df_all, df_donor)
}
\arguments{
\item{village}{A Village object.}

\item{lowci}{Lower bound of credible interval}

\item{highci}{Higher bound of credible interval}

\item{df_all}{Optional data frame containing treatment parameter draws.}

\item{df_donor}{Data frame with distinct donor ids/names.}
}
\value{
None
}
\description{
Internal generic function to plot violin of total donor proliferation effect per treatment dose.
}
\keyword{internal}
