% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_growthmetric}
\alias{plot_growthmetric}
\title{Plot total proliferation effect}
\usage{
plot_growthmetric(
  village,
  lowci,
  highci,
  df_treat = NULL,
  df_cont,
  df_donor,
  df_dcov,
  returnboth = FALSE
)
}
\arguments{
\item{village}{A Village object.}

\item{lowci}{Lower bound of credible interval}

\item{highci}{Higher bound of credible interval}

\item{df_treat}{Optional data frame containing treatment parameter draws.}

\item{df_cont}{Data frame containing control intercept draws.}

\item{df_donor}{Data frame with distinct donor ids/names.}

\item{df_dcov}{Donor covariate parameter draws.}

\item{returnboth}{If TRUE returns list of data and plots. Data includes lfsr calculations.}
}
\value{
List of data (lfsr) and plot.
}
\description{
Internal generic function to plot violin of total donor proliferation effect per treatment dose.
}
\keyword{internal}
