% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_overdispersion}
\alias{plot_overdispersion}
\title{Plot overdispersion parameters}
\usage{
plot_overdispersion(village, lowci, highci)
}
\arguments{
\item{village}{A Village object.}

\item{lowci}{Lower bound of credible interval}

\item{highci}{Higher bound of credible interval}
}
\value{
None
}
\description{
Internal generic function to plot violin of theta, phi, and phi_r.
}
\keyword{internal}
