% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_params}
\alias{plot_params}
\title{Plot all model parameters and growth rates}
\usage{
plot_params(village)
}
\arguments{
\item{village}{A Village object.}
}
\value{
An updated village object. Also saves a diagnostic plot.
}
\description{
Internal generic function to plot proliferation and dispersion model parameters.
}
\keyword{internal}
