% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_ppck}
\alias{plot_ppck}
\title{Plot posterior predictive checks}
\usage{
plot_ppck(village)
}
\arguments{
\item{village}{A Village object.}
}
\value{
An updated village object. Also saves a diagnostic plot.
}
\description{
Internal generic function to plot posterior predictive checks.
}
\keyword{internal}
