% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_rawdata}
\alias{plot_rawdata}
\title{Plot raw data}
\usage{
plot_rawdata(village)
}
\arguments{
\item{village}{A Village object.}
}
\value{
None.
}
\description{
Internal generic function to plot scatterplot and stacked bar plot of raw representation data.
}
\keyword{internal}
