% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_townlet.R
\name{sumstats}
\alias{sumstats}
\title{Model summary statistics}
\usage{
sumstats(village)
}
\arguments{
\item{village}{A Village object.}
}
\value{
An updated village object.
}
\description{
Internal generic function that generates model parameters summary statistics.
}
\keyword{internal}
