% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-townlet.R
\name{test_townlet}
\alias{test_townlet}
\title{Check Package Installation}
\usage{
test_townlet()
}
\value{
Fitted model results from test data
}
\description{
Runs test model to verify that Townlet work correctly after installation.
}
