% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_village.R
\name{validate_village}
\alias{validate_village}
\title{Checks data and model structure}
\usage{
validate_village(village)
}
\arguments{
\item{village}{A Village object.}
}
\value{
An updated village object. Also saves a diagnostic plot.
}
\description{
Internal generic function that checks that all init_village() inputs are in correct format.
}
\keyword{internal}
