% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_village.R
\name{Village}
\alias{Village}
\title{Helper function that initiates village object}
\usage{
Village(
  datapath,
  model,
  normalize,
  alldonors,
  T0_cutoff,
  name,
  timeunit,
  baseline,
  colors,
  sim,
  ebayes,
  outdir
)
}
\description{
Helper function that initiates village object
}
\keyword{internal}
