% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_townlet.R
\name{run_townlet}
\alias{run_townlet}
\title{Run townlet donor proliferation analysis}
\usage{
run_townlet(
  village,
  cores = 1,
  samples = 2000,
  warmup = 1000,
  chains = 4,
  credinterval = 0.95,
  ppck = TRUE,
  priors = NULL,
  comp = 1
)
}
\arguments{
\item{village}{Village object generated by townlet::init_village()}

\item{cores}{Number of cores for sampling}

\item{samples}{Number of samples (including warmup)}

\item{warmup}{Number of warmup samples}

\item{chains}{Number of chains}

\item{credinterval}{Credible interval for significance testing}

\item{ppck}{Run posterior predictive check}

\item{priors}{Set custom priors}

\item{comp}{If running on server comp=1, else if running on local computer comp=0}
}
\value{
Updated village object.
}
\description{
Runs hierarchical Dirichlet regression model to quantify treatment specific donor proliferation effects and test whether specific donor groups significantly influence growth.
}
\examples{
\dontrun{
  test_village <- run_townlet(test_village)
}
}
