% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_village.R
\name{t0_checkrep}
\alias{t0_checkrep}
\title{Check donor time-0 representation}
\usage{
t0_checkrep(village)
}
\arguments{
\item{village}{A Village object.}
}
\value{
An updated village object. Also saves a diagnostic plot.
}
\description{
Internal generic function to check donor representation at time 0.
}
\keyword{internal}
