% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_village.R
\name{Village}
\alias{Village}
\title{Helper function that initiates village object}
\usage{
Village(
  datapath,
  model,
  normalize = FALSE,
  outdir = "./",
  name = "testmodel",
  cls,
  T0_cutoff,
  alldonors = TRUE,
  sim = FALSE,
  timeunit,
  baseline = NULL,
  ebayes
)
}
\description{
Helper function that initiates village object
}
\keyword{internal}
